GEFIXED!!!

+++ Code/CNC_Interpreter_V2/GPIOControl.cs
@@ -136,7 +136,7 @@ namespace CNC_Interpreter_V2
                 Speed = 0;
             }
             Console.WriteLine("DutyCycle Calculation: " + Speed.ToString() + " / 100 = " + ((double)Speed / 100).ToString());
-            double DutyCycle = Speed / 100;
+            double DutyCycle = (double)Speed / 100;
             Console.WriteLine("DC: " + DutyCycle.ToString());
             _setPWM(true, 1, 0, DutyCycle);
             _setPin(6, Dir);
